@@ -1,13 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
+import Button from '@mui/material/Button';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogTitle from '@mui/material/DialogTitle';
 
-import I18n from '@iobroker/adapter-react/i18n';
+import I18n from '@iobroker/adapter-react-v5/i18n';
 
 const CustomModal = ({ open, close, children, titleButton, titleButton2, buttonClick }) => {
     return <Dialog
@@ -22,7 +22,7 @@ const CustomModal = ({ open, close, children, titleButton, titleButton2, buttonC
             {children}
         </DialogContent>
         <DialogActions>
-            {titleButton && <Button variant="contained" onClick={buttonClick}>
+            {titleButton && <Button color="grey" variant="contained" onClick={buttonClick}>
                 {titleButton}
             </Button>}
             {titleButton2 && <Button variant="contained" onClick={close} color="primary" autoFocus>
