@@ -1,43 +1,33 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import * as Sentry from '@sentry/browser';
-import * as SentryIntegrations from '@sentry/integrations';
-import { MuiThemeProvider} from '@material-ui/core/styles';
+import { ThemeProvider, StyledEngineProvider } from '@mui/material/styles';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import {version} from '../package.json';
-import theme from '@iobroker/adapter-react/Theme';
-import Utils from '@iobroker/adapter-react/Components/Utils';
+import pkg from '../package.json';
+import theme from '@iobroker/adapter-react-v5/Theme';
+import Utils from '@iobroker/adapter-react-v5/Components/Utils';
 
 window.adapterName = 'web';
+window.sentryDSN = 'https://5ad729dbed504d15aa8bde423cae9a8e@sentry.iobroker.net/57';
 let themeName = Utils.getThemeName();
 
-
-console.log('iobroker.' + window.adapterName + '@' + version + ' using theme "' + themeName + '"');
+console.log('iobroker.' + window.adapterName + '@' + pkg.version + ' using theme "' + themeName + '"');
 
 function build() {
     return ReactDOM.render(
-        <MuiThemeProvider theme={theme(themeName)}>
-            <App
-                onThemeChange={_theme => {
-                    themeName = _theme;
-                    build();
-                }}
-            />
-        </MuiThemeProvider>,
+        <StyledEngineProvider injectFirst>
+            <ThemeProvider theme={theme(themeName)}>
+                <App
+                    onThemeChange={_theme => {
+                        themeName = _theme;
+                        build();
+                    }}
+                />
+            </ThemeProvider>
+        </StyledEngineProvider>,
         document.getElementById('root')
     );
-}
-
-if (window.location.host !== 'localhost:3000') {
-    Sentry.init({
-        dsn: 'https://5ad729dbed504d15aa8bde423cae9a8e@sentry.iobroker.net/57',
-        release: 'iobroker.' + window.adapterName + '@' + version,
-        integrations: [
-            new SentryIntegrations.Dedupe()
-        ]
-    });
 }
 
 build();
